GT-1 - Extracted SubscriptionRedirect
