+++ app/controllers/UIModel.scala
@@ -97,13 +97,3 @@ object AjaxSubscriptionResponse{
     AjaxSubscriptionResponse(context, group.toString, subscriptionNumberLabel(apiSubscriptions))
   }
 }
-
-sealed trait SubscriptionRedirect extends EnumEntry
-
-object SubscriptionRedirect extends PlayEnum[SubscriptionRedirect] {
-  val values = findValues
-
-  final case object MANAGE_PAGE               extends SubscriptionRedirect
-  final case object APPLICATION_CHECK_PAGE    extends SubscriptionRedirect
-  final case object API_SUBSCRIPTIONS_PAGE    extends SubscriptionRedirect
-}
