+++ app/views/include/subscribeConfirmation.scala.html
@@ -15,7 +15,7 @@
  *@
 
 @import config.ApplicationConfig
-@import domain.{Application, Developer}
+@import domain.{Application, Developer, SubscriptionRedirect}
 @import model.Crumb
 @import include._
