+++ app/views/include/subscriptionGroups.scala.html
@@ -14,7 +14,7 @@
  * limitations under the License.
  *@
 
-@import domain.{Application, Developer, Role}
+@import domain.{Application, Developer, Role, SubscriptionRedirect}
 @import include._
 
 @(role: Role, subscriptions: GroupedSubscriptions, app: Application, hasSubscriptions: Boolean, afterUnsubscribeRedirectTo: SubscriptionRedirect)(
