+++ app/views/include/subscriptionsGroup.scala.html
@@ -15,7 +15,7 @@
  *@
 
 @import config.ApplicationConfig
-@import domain.{APISubscriptionStatus, Application, Environment, Role, State}
+@import domain.{APISubscriptionStatus, Application, Environment, Role, SubscriptionRedirect, State}
 @import uk.gov.hmrc.urls.Link
 @import views.helper.IdFormatter
